requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
flake8==7.3.0
flake8-broken-line==1.0.0
flake8-docstrings==1.7.0
flake8-isort==7.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
idna==3.7
isort==7.0.0
mccabe==0.7.0
pep8-naming==0.15.1
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.3
scratchattach==2.1.13
setuptools==81.0.0
SimpleWebSocketServer==0.1.2
snowballstemmer==3.0.1
soupsieve==2.8.3
srequests==0.0.3
typing_extensions==4.15.0
urllib3==2.2.2
websocket-client==1.9.0
